update: исправлена логика работы

+++ main-server/src/main/java/ru/practicum/app/compilation/CompilationServiceImpl.java
@@ -1,13 +1,15 @@
 package ru.practicum.app.compilation;
 
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ru.practicum.app.event.EventRepository;
 import ru.practicum.app.event.EventServiceImpl;
 import ru.practicum.app.event.EventShortDto;
 
-import javax.persistence.EntityNotFoundException;
+import javax.persistence.*;
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
 import java.util.List;
@@ -19,6 +21,8 @@ import java.util.stream.Collectors;
 public class CompilationServiceImpl {
     private final CompilationRepository compilationRepository;
     private final CompilationEventRepository compilationEventRepository;
+
+    private final EventRepository eventRepository;
     private final EventServiceImpl eventService;
 
     private final CompilationMapper compilationMapper;
@@ -26,11 +30,13 @@ public class CompilationServiceImpl {
     @Autowired
     public CompilationServiceImpl(CompilationRepository compilationRepository,
                                   CompilationEventRepository compilationEventRepository,
-                                  EventServiceImpl eventService, CompilationMapper compilationMapper) {
+                                  EventServiceImpl eventService, CompilationMapper compilationMapper,
+                                  EventRepository eventRepository) {
         this.compilationRepository = compilationRepository;
         this.compilationEventRepository = compilationEventRepository;
         this.eventService = eventService;
         this.compilationMapper = compilationMapper;
+        this.eventRepository = eventRepository;
     }
 
     public List<CompilationDto> getCompilations(Boolean pinned, @Min(0) int from, @Min(1) int size) {
@@ -46,24 +52,74 @@ public class CompilationServiceImpl {
         return CompilationMapper.mapToCompilationDto(compilation, getCompilationEvents(compId));
     }
 
+//    public class NewCompilationDto {
+//        private String title;
+//        private Boolean pinned;
+//        private List<Integer> events;
+
+//    @Entity
+//    @Table(name = "compilation_event")
+//    @Data
+//    @NoArgsConstructor
+//    @AllArgsConstructor
+//    public class CompilationEvent {
+//        @Id
+//        @GeneratedValue(strategy = GenerationType.IDENTITY)
+//        private Integer id;
+//        @Column(name = "compilation_id")
+//        private Integer compilation;
+//        @Column(name = "event_id")
+//        private Integer event;
+//        @Column(name = "title", nullable = false)
+//        String title;
+//        @Column(name = "pinned", nullable = false)
+//        Boolean pinned;
+
+
     public CompilationDto addCompilation(@Valid NewCompilationDto newCompilationDto) {
+//
+//        List<Event> events = new ArrayList<>();
+//
+//        for (Integer eventId : newCompilationDto.getEvents()) {
+//            events.add(eventRepository.findById(eventId).get());
+//        }
+//
+//        Compilation compilation = new Compilation();
+//        compilation.setTitle(newCompilationDto.getTitle());
+//        compilation.setPinned(newCompilationDto.getPinned());
+//        compilationRepository.save(compilation);
+//
+//        for (Event eventId : events) {
+//            CompilationEvent compilationEvent = new CompilationEvent(null, compId, eventId, null, null);
+//            compilationEventRepository.save(compilationEvent);
+//        }
+//
+//
+//        return CompilationMapper.toCompilationDto(compilation);
+
+
         Compilation compilation = compilationMapper.mapToCompilation(newCompilationDto);
         Compilation addedCompilation = compilationRepository.save(compilation);
         Integer compId = addedCompilation.getId();
 
         for (Integer eventId : newCompilationDto.getEvents()) {
-            CompilationEvent compilationEvent = new CompilationEvent(null, compId, eventId, null, null);
+            CompilationEvent compilationEvent = new CompilationEvent(null, compId, eventId);
             compilationEventRepository.save(compilationEvent);
         }
         return CompilationMapper.mapToCompilationDto(addedCompilation, getCompilationEvents(compId));
     }
 
     public void deleteCompilationById(Integer compId) {
+        if (compId == null) {
+            throw new CompilationNotFoundException("подборка не найдена");
+        }
         compilationRepository.deleteById(compId);
+//        compilationEventRepository.deleteById(compId);
+
     }
 
     public void addEventToCompilation(Integer compId, Integer eventId) {
-        CompilationEvent compilationEvent = new CompilationEvent(null, compId, eventId, null, null);
+        CompilationEvent compilationEvent = new CompilationEvent(null, compId, eventId);
         compilationEventRepository.save(compilationEvent);
     }
